backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
celery[redis]==5.4.0
redis==5.2.0
websockets==13.1
python-multipart==0.0.12
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
pydantic==2.10.1
pydantic-settings==2.6.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
flower==2.0.1
stripe==11.1.0
resend==2.4.0
anthropic==0.39.0
requests==2.32.3
python-dateutil==2.9.0
